+++ src/Controller/API/AdminStatsController.php
@@ -2,8 +2,10 @@
 
 namespace App\Controller\API;
 
+use App\Entity\Projet;
 use App\Entity\User;
 use App\Repository\CraRepository;
+use App\Repository\TempsPasseRepository;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -68,6 +70,57 @@ class AdminStatsController extends AbstractController
 
         return new JsonResponse([
             'months' => $data,
+        ]);
+    }
+
+    /**
+     * Retourne les temps passés sur un projet par les contributeurs sur une année.
+     * Exemple :
+     * {
+     *      "months": [
+     *          {
+     *              "User A": 35,
+     *              "User B": 5
+     *          },
+     *          {
+     *              "User A": 40,
+     *              "User C": 2
+     *          },
+     *          ...
+     *      ]
+     * }
+     *
+     * @Route(
+     *      "/temps-par-user/{id}/{year}",
+     *      methods={"GET"},
+     *      requirements={"year"="\d{4}"},
+     *      name="api_stats_admin_temps_projet_users"
+     * )
+     */
+    public function getTempsProjetParUsers(
+        string $year,
+        Projet $projet,
+        TempsPasseRepository $tempsPasseRepository
+    ) {
+        $this->denyAccessUnlessGranted('same_societe', $projet);
+
+        $tempsPasse = $tempsPasseRepository->findAllByProjetAndYear($projet, $year);
+        $months = [];
+
+        for ($i = 0; $i < 12; ++$i) {
+            $months[$i] = [];
+        }
+
+        foreach ($tempsPasse as $tempsPasse) {
+            $month = intval($tempsPasse->getCra()->getMois()->format('m')) - 1;
+            $user = $tempsPasse->getCra()->getUser()->getFullname();
+            $percentage = $tempsPasse->getPourcentage();
+
+            $months[$month][$user] = $percentage;
+        }
+
+        return new JsonResponse([
+            'months' => $months,
         ]);
     }
 }